Some Object methods should be implemented on Kernel level
